+++ lib/Models/WebMapServiceCatalogItem.ts
@@ -446,19 +446,13 @@ class DiffStratum extends LoadableStratum(WebMapServiceCatalogItemTraits) {
     const firstDate = this.catalogItem.firstDiffDate;
     const secondDate = this.catalogItem.secondDiffDate;
     if (diffStyleId && firstDate && secondDate) {
-      return this.catalogItem.getLegendUrlForDiffStyle(
+      return this.catalogItem.getLegendUrlForStyle(
         diffStyleId,
         JulianDate.fromIso8601(firstDate),
         JulianDate.fromIso8601(secondDate)
       );
     }
     return undefined;
-  }
-
-  @computed
-  get availableDiffStyles() {
-    // Currently only NDVI
-    return ["NDVI"];
   }
 
   @computed
@@ -623,6 +617,14 @@ class WebMapServiceCatalogItem
     }
   }
 
+  @computed
+  get canDiffImages(): boolean {
+    const hasValidDiffStyles = this.availableDiffStyles.some(diffStyle =>
+      this.styleSelector?.availableStyles.find(style => style.id === diffStyle)
+    );
+    return hasValidDiffStyles === true;
+  }
+
   showDiffImage(
     firstDate: JulianDate,
     secondDate: JulianDate,
@@ -648,24 +650,26 @@ class WebMapServiceCatalogItem
     this.setTrait(CommonStrata.user, "isShowingDiff", false);
   }
 
-  getLegendUrlForDiffStyle(
-    diffStyleId: string,
-    firstDate: JulianDate,
-    secondDate: JulianDate
+  getLegendUrlForStyle(
+    styleId: string,
+    firstDate?: JulianDate,
+    secondDate?: JulianDate
   ) {
-    const firstTag = this.getTagForTime(firstDate);
-    const secondTag = this.getTagForTime(secondDate);
-    const time = `${firstTag},${secondTag}`;
+    const firstTag = firstDate && this.getTagForTime(firstDate);
+    const secondTag = secondDate && this.getTagForTime(secondDate);
+    const time = filterOutUndefined([firstTag, secondTag]).join(",");
     const layerName = this.availableStyles.find(style =>
-      style.styles.some(s => s.name === diffStyleId)
+      style.styles.some(s => s.name === styleId)
     )?.layerName;
-    return URI(
+    const uri = URI(
       `${this.url}?service=WMS&version=1.1.0&request=GetLegendGraphic&format=image/png&transparent=True`
     )
       .addQuery("layer", encodeURIComponent(layerName || ""))
-      .addQuery("styles", encodeURIComponent(diffStyleId))
-      .addQuery("time", time)
-      .toString();
+      .addQuery("styles", encodeURIComponent(styleId));
+    if (time) {
+      uri.addQuery("time", time);
+    }
+    return uri.toString();
   }
 
   @computed