+++ CHANGES.md
@@ -12,6 +12,7 @@ Change Log
 * Added file upload support for `GltfCatalogItem`. The supported extension is glb.
 * Improve runtime themeing via styled components across main UI components
 * Updated default welcome video defaults to a newer, slower video
+* Difftool will now pick any existing marked location (like from a search result) and filter imagery for that location.
 * ChartCustomComponent now `abstract`, no longer specific to CSV catalog items. Implement it for custom feature info charts.
 * (💫The next rad feature💫)
 * Update date picker to use theme colours